Rename jd-cd to jd-path
It will only ever be possibly to use jd-cd to display the path to a
johnny.decimal directory, as I just learned that changing the working
directory of the parent process should not and can not be easily done by
a subprocess.

It makes sense, therefore, to rename jd-cd to jd-path. Instructions on
how to integrate jd-path into the shell and use it to *actually* perform
a cd-like operation will be added to the jd-path manual page in the
future.