Fix optind related error messages
If a short option with an argument is specified without a space between
them, like

jd config -ujd_root

getopt(3) doesn't increase "optind" to the next argv element, which, on
error, results in "argv[optind - 1]" not pointing the start of the
violating argv element, but one prior to it.

It makes sense, therefore, to check whether "optind" has changed since
the last iteration and adjust the error message accordingly.